pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yukon-mcp-service"
version = "0.1.0"
description = "Yukon MCP Server using fastMCP framework"
authors = [
    {name = "Zhang Ming", email = "zhangming1@supermap.com"}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["mcp", "yukon","postgis", "spatial", "gis"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcp>=0.1.0",
    "psycopg2-binary>=2.9.0",
    "asyncpg>=0.29.0",
    "shapely>=2.0.0",
    "geopandas>=0.14.0",
    "fiona>=1.9.0",
    "pyproj>=3.6.0",
    "rasterio>=1.3.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "vanna>=0.5.0",
    "chromadb>=0.4.0",
    "openai>=1.0.0",
    "anthropic>=0.7.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic[email]>=2.0.0",
    "python-multipart>=0.0.6",
    "flask>=3.0.0",
    "fastmcp>=2.13.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[project.scripts]
yukon-mcp = "src.server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
